#530 Started adding some tests around how flowchart shapes are rendered in SVG
